pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "shape-canvas"
version = "1.0.0"
description = "A powerful Python library for drawing geometric shapes and decorative elements on canvas using PIL/Pillow"
readme = "README.md"
license = "MIT"
authors = ["K. Dasaradha <kdasaradha525@example.com>"]
maintainers = ["K. Dasaradha <kdasaradha525@example.com>"]
homepage = "https://github.com/KDASARADHA525/Shapes-python-pillow-canvas"
repository = "https://github.com/KDASARADHA525/Shapes-python-pillow-canvas.git"
documentation = "https://github.com/KDASARADHA525/Shapes-python-pillow-canvas#readme"
keywords = ["graphics", "shapes", "canvas", "PIL", "Pillow", "drawing", "2D"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "shape_canvas", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8.1"
Pillow = ">=10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.0.0"
black = ">=23.0.0"
flake8 = ">=6.0.0"
mypy = ">=1.0.0"

[tool.poetry.scripts]
shape-canvas = "shape_canvas.cli:main"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/KDASARADHA525/Shapes-python-pillow-canvas/issues"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true